iox-#64 Move Channel data structure into its own file.
